add src/options.ts
